+++ example/extending/newlang/6-python-modules/qux/src/main.py
@@ -6,10 +6,10 @@ def main():
 
     # Check if any arguments were provided
     if not args:
-        print("Hello, World!")
+        print("Hello, World Qux!")
     else:
         # Join the arguments with spaces and print the greeting
-        greeting = "Hello, " + " ".join(args) + "!"
+        greeting = "Hello, " + " ".join(args) + " Qux!"
         print(greeting)
 
 if __name__ == "__main__":
