+++ example/extending/newlang/7-python-deps/foo/bar/src/requirements.txt
@@ -1,2 +1,1 @@
-numpy
 pandas
