+++ example/extending/newlang/7-python-deps/foo/src/requirements.txt
@@ -1,0 +1,1 @@
+numpy
