+++ example/extending/newlang/9-python-pex/build.mill
@@ -23,11 +23,10 @@ import mill._
  */
 trait PythonModule extends Module {
 
+  def moduleDeps: Seq[PythonModule] = Nil
+
   /** Specifies dependencies to install in the virtual environment. */
   def dependencies: T[Seq[String]] = Task { Seq.empty[String] }
-
-  /** Specifies command-line options to pass to the main Python file. */
-  def options: T[Seq[String]] = Task { Seq.empty[String] }
 
   /** Specifies the name of the virtual environment directory. */
   def venvName: T[String] = Task { "venv" }
@@ -110,20 +109,23 @@ trait PythonModule extends Module {
     os.call(Seq("bash", "-c", s"source ${venvBin / "activate"}"))
     PathRef(venvBin)
   }
+  
+  def finalDependencies: T[Seq[String]] = Task {
+    val upstreamDependencies = Task.traverse(moduleDeps)(_.finalDependencies)()
+    (dependencies().map(_.toString) ++ requirementsDependencies().map(_.toString) ++ upstreamDependencies.flatten.map(_.toString)).distinct
+  }
 
   /** Installs specified dependencies in the virtual environment. */
   def installDependencies: T[PathRef] = Task {
     val pythonVenv = activateVenv().path / "python3"
 
-    val pythonDependencies: Seq[String] = dependencies() ++ requirementsDependencies()
-
-    if (pythonDependencies.nonEmpty) {
+    if (finalDependencies().nonEmpty) {
       os.call(Seq(
         pythonVenv.toString,
         "-m",
         "pip",
         "install"
-      ) ++ pythonDependencies ++ Seq(
+      ) ++ finalDependencies() ++ Seq(
         "--break-system-packages",
         "--no-warn-script-location"
       ))
@@ -132,62 +134,68 @@ trait PythonModule extends Module {
     PathRef(pythonVenv)
   }
 
-  /** Retrieves the path to the main Python file, ensuring it exists. */
-  def mainFile: T[PathRef] = Task {
-
-    val mainFilePath = allSources().find(_.path.last == mainFileName()) match {
-      case Some(mainFile) => mainFile.path
-      case None =>
-        throw new java.io.FileNotFoundException(
-          s"Main file '${mainFileName()}' not found in sources."
-        )
-    }
-    PathRef(mainFilePath)
-  }
-
   /** Executes the main Python file using the virtual environment. */
-  def compile: T[Unit] = Task {
+  def compile: T[(PathRef, PathRef)] = Task {
     val pythonVenv = installDependencies().path
-    val mainFilePath = mainFile().path
-    val args = options()
-
-    val result = os.call(Seq(pythonVenv.toString, mainFilePath.toString) ++ args)
-
-    T.log.info(result.out.text())
-    if (result.exitCode != 0) {
-      T.log.error(s"Error: ${result.err.text()}")
-    }
+    val mainFileDir = sources().path
+    
+    (PathRef(pythonVenv), PathRef(mainFileDir))
   }
-
-  /** Executes the main Python file with pex Library using the virtual environment. */
-  def compilePex: T[PathRef] = Task {
-    val pythonVenv = installDependencies().path
-    val mainFilePath = mainFile().path
-    val args = options()
+  
+  def run(args: mill.define.Args) = Task.Command {
+    
     val pexFile = T.dest / "result.pex"
 
-    val result = os.call(
-      Seq(
-        pythonVenv.toString,
-        "-m",
-        "pex",
-        "-r",
-        requirementsPath().path.toString,
-        "-c",
-        mainFilePath.toString,
-        "-o",
-        pexFile.toString
-      ) ++ args
-    )
-
-    println(result)
-    T.log.info(result.out.text())
-    if (result.exitCode != 0) {
-      T.log.error(s"Error: ${result.err.text()}")
+    val upstream = Task.traverse(moduleDeps)(_.compile)().zip(moduleDeps)
+    for(((_, mainFileDir), mod) <- upstream) {
+      val destinationPath = Task.dest / mod.millSourcePath.subRelativeTo(build.millSourcePath)
+      os.copy.over(mainFileDir.path / os.up, destinationPath)
     }
 
-    PathRef(pexFile)
+    os.call(
+      Seq(
+        compile()._1.path.toString,
+        "-m",
+        "pex",
+      ) ++ finalDependencies() ++ Seq(
+        "-c",
+        s"${compile()._2.path / mainFileName()}",
+        "-o",
+        pexFile.toString
+      ) ++ args.value,
+      env = Map("PYTHONPATH" -> Task.dest.toString)
+    )
   }
+
+  // /** Executes the main Python file with pex Library using the virtual environment. */
+  // def compilePex: T[PathRef] = Task {
+  //   val pythonVenv = installDependencies().path
+  //   val mainFilePath = mainFile().path
+  //   val args = options()
+  //   val pexFile = T.dest / "result.pex"
+
+  //   val result = os.call(
+  //     Seq(
+  //       pythonVenv.toString,
+  //       "-m",
+  //       "pex",
+  //       "-r",
+  //       requirementsPath().path.toString,
+  //       "-c",
+  //       mainFilePath.toString,
+  //       "-o",
+  //       pexFile.toString
+  //     ) ++ args
+  //   )
+
+  //   println(result)
+  //   T.log.info(result.out.text())
+  //   if (result.exitCode != 0) {
+  //     T.log.error(s"Error: ${result.err.text()}")
+  //   }
+
+  //   PathRef(pexFile)
+  // }
 
 }
 
@@ -211,7 +219,7 @@ object pex extends PythonModule{
 
 /** Usage
 
-> ./mill show pex.compilePex
+> ./mill show pex.run
 ".../out/pex/compilePex.dest/result.pex"
 
 */
